Moved SortableNameConverter to its own file
