+++ source/Playnite/Playnite.csproj
@@ -227,6 +227,7 @@
     <Compile Include="Behaviors\SelectorBehaviors.cs" />
     <Compile Include="Common\CueSheet.cs" />
     <Compile Include="Common\Extensions\KeyExtensions.cs" />
+    <Compile Include="SortableNameConverter.cs" />
     <Compile Include="Common\FileSystem_Checksum.cs" />
     <Compile Include="Controls\ExtendedDataGrid.cs" />
     <Compile Include="ControlTemplateTools.cs" />
