+++ source/Playnite/SortableNameConverter.cs
@@ -1,0 +1,209 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+
+namespace Playnite
+{
+    public class SortableNameConverter
+    {
+        public IEnumerable<string> Articles { get; }
+
+        private Regex regex;
+
+        /// <summary>
+        /// The minimum string length of numbers. If 4, XXIII or 23 will turn into 0023.
+        /// </summary>
+        private static int NumberLength = 2;
+
+        private static string[] ExcludedRomanNumerals = new[] { "XL", "XD", "XXX", "D", "MII", "MIX", "MX", "MC" };
+
+        //When adding/removing entries here, be sure to update the regex too
+        private static Dictionary<string, int> NumberWordValues = new Dictionary<string, int>(StringComparer.InvariantCultureIgnoreCase) { { "one", 1 }, { "two", 2 }, { "three", 3 } };
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="articles">Words to remove from the start of the title. Suggested: the contents of PlayniteSettings.GameSortingNameRemovedArticles, or "The", "A", "An".</param>
+        /// <param name="batchOperation">Optimize for larger amounts of throughput. Slower for small amounts.</param>
+        public SortableNameConverter(IEnumerable<string> articles, bool batchOperation = false)
+        {
+            Articles = articles ?? throw new ArgumentNullException(nameof(articles));
+            string articlesPattern = string.Join("|", articles.Select(Regex.Escape));
+            var options = RegexOptions.ExplicitCapture;
+            if (batchOperation)
+                options |= RegexOptions.Compiled;
+
+            //(?<![\w.]|^) prevents the numerical matches from happening at the start of the string (for example for X-COM or XIII) or attached to a word or . (to avoid S.T.A.L.K.E.R. -> S.T.A.50.K.E.R.)
+            //(?!\.) prevents matching roman numerals with a period right after (again for cases like abbreviations with periods, but that start with a roman numeral character)
+            //\u2160-\u2188 is the unicode range of roman numerals listed in RomanNumeralValues
+            //using [0-9] here instead of \d because \d also matches ٠١٢٣٤٥٦٧٨٩ and I don't know what to do with those           
+            //the (?i) is a modifier that makes the rest of the regex (to the right of it) case insensitive
+            //see https://www.regular-expressions.info/modifiers.html
+            regex = new Regex($@"(?<![\w.]|^)((?<roman>[IVXLCDM\u2160-\u2188]+(?!\.))|(?<arabic>[0-9]+))(?=\W|$)|(?i)^(?<article>{articlesPattern})\s+|\b(?<numberword>one|two|three)\b", options);
+        }
+
+        public string Convert(string input)
+        {
+            if (string.IsNullOrWhiteSpace(input))
+            {
+                return input;
+            }
+
+            return regex.Replace(input, match =>
+            {
+                if (match.Groups["roman"].Success)
+                {
+                    if (match.Value == "I")
+                    {
+                        if (MatchComesAfterChapterOrEpisodeOrAtEndOfString(input, match))
+                        {
+                            return "1".PadLeft(NumberLength, '0');
+                        }
+                        else
+                        {
+                            return match.Value;
+                        }
+                    }
+                    else if (ExcludedRomanNumerals.Contains(match.Value))
+                    {
+                        return match.Value;
+                    }
+                    return ConvertRomanNumeralToInt(match.Value)?.ToString(new string('0', NumberLength)) ?? match.Value;
+                }
+                else if (match.Groups["arabic"].Success)
+                {
+                    return match.Value.PadLeft(NumberLength, '0');
+                }
+                else if (match.Groups["article"].Success)
+                {
+                    return string.Empty;
+                }
+                else if (match.Groups["numberword"].Success)
+                {
+                    if (MatchComesAfterChapterOrEpisodeOrAtEndOfString(input, match))
+                    {
+                        return NumberWordValues[match.Value].ToString(new string('0', NumberLength));
+                    }
+                    else
+                    {
+                        return match.Value;
+                    }
+                }
+                return match.Value;
+            });
+        }
+
+        private static bool MatchComesAfterChapterOrEpisodeOrAtEndOfString(string input, Match match)
+        {
+            bool matchIsAtEndOfString = match.Index + match.Length == input.Length;
+            string theBitImmediatelyPriorToTheMatch = input.Substring(Math.Max(0, match.Index - 9), length: Math.Min(9, match.Index));
+            return matchIsAtEndOfString
+                || theBitImmediatelyPriorToTheMatch.Contains("chapter", StringComparison.InvariantCultureIgnoreCase)
+                || theBitImmediatelyPriorToTheMatch.Contains("episode", StringComparison.InvariantCultureIgnoreCase);
+        }
+
+        private static Dictionary<char, int> RomanNumeralValues = new Dictionary<char, int>
+        {
+            { 'I', 1 }, { 'V', 5 }, { 'X', 10 }, { 'L', 50 }, { 'C', 100 }, { 'D', 500 }, { 'M', 1000 },
+            //unicode uppercase
+            {'Ⅰ', 1}, {'Ⅱ', 2}, {'Ⅲ', 3}, {'Ⅳ', 4}, {'Ⅴ', 5}, {'Ⅵ', 6}, {'Ⅶ', 7}, {'Ⅷ', 8}, {'Ⅸ', 9}, {'Ⅹ', 10}, {'Ⅺ', 11}, {'Ⅻ', 12}, {'Ⅼ', 50}, {'Ⅽ', 100}, {'Ⅾ', 500}, {'Ⅿ', 1000},
+            //unicode lowercase
+            {'ⅰ', 1}, {'ⅱ', 2}, {'ⅲ', 3}, {'ⅳ', 4}, {'ⅴ', 5}, {'ⅵ', 6}, {'ⅶ', 7}, {'ⅷ', 8}, {'ⅸ', 9}, {'ⅹ', 10}, {'ⅺ', 11}, {'ⅻ', 12}, {'ⅼ', 50}, {'ⅽ', 100}, {'ⅾ', 500}, {'ⅿ', 1000},
+            //unicode big/exotic numbers
+            {'ↀ', 1000}, {'ↁ', 5000}, {'ↂ', 10000}, {'Ↄ', 100}, {'ↄ', 100}, {'ↅ', 6}, {'ↆ', 50 }, {'ↇ', 50000}, {'ↈ', 100000 }
+        };
+
+        /// <summary>
+        /// Convert a number from Roman numerals to an integer
+        /// </summary>
+        /// <param name="input">The roman numeral.</param>
+        /// <param name="validate">If false, parse any roman numeral. If true, reject invalid ones.</param>
+        /// <returns>An integer form of the supplied roman numeral, or NULL if the supplied roman numeral is invalid and <paramref name="validate"/> is true.</returns>
+        /// <exception cref="KeyNotFoundException">When the input contains non-numeral characters.</exception>
+        public static int? ConvertRomanNumeralToInt(string input, bool validate = true)
+        {
+            if (string.IsNullOrWhiteSpace(input))
+            {
+                return null;
+            }
+
+            int output = 0;
+            int biggestNumberToTheRight = 0;
+
+            int prevCharGroupLength = 0;
+            int lastNumericValue = 0;
+            for (int i = input.Length - 1; i >= 0; i--)
+            {
+                char c = input[i];
+                int value = RomanNumeralValues[c];
+                bool subtract = value < biggestNumberToTheRight;
+                if (subtract)
+                {
+                    output -= value;
+                }
+                else
+                {
+                    output += value;
+                    biggestNumberToTheRight = value;
+                }
+
+                #region validation
+
+                if (!validate)
+                {
+                    continue;
+                }
+
+                //reject things like IVX and VIX and IIX
+                //subtractive numerals are only ever singular
+                if (subtract && lastNumericValue < biggestNumberToTheRight)
+                {
+                    return null;
+                }
+
+                //reject things like VX or LC or DM
+                //subtractions can't be half the bigger value
+                //IV is as close as the two numbers get in value
+                if (subtract && value * 5 > biggestNumberToTheRight)
+                {
+                    return null;
+                }
+
+                if (value == lastNumericValue)
+                {
+                    //Numerals that aren't 1 or 10ⁿ can't repeat
+                    if (!IsOneOrPowerOfTen(value))
+                    {
+                        return null;
+                    }
+
+                    //No numeral can repeat 4 times
+                    prevCharGroupLength++;
+                    if (prevCharGroupLength == 4)
+                    {
+                        return null;
+                    }
+                }
+                else
+                {
+                    prevCharGroupLength = 1;
+                }
+
+                lastNumericValue = value;
+
+                #endregion validation
+            }
+            return output;
+        }
+
+        private static bool IsOneOrPowerOfTen(int x)
+        {
+            while (x > 9 && x % 10 == 0)
+            {
+                x /= 10;
+            }
+            return x == 1;
+        }
+    }
+}
