+++ source/Tests/Playnite.Tests/Playnite.Tests.csproj
@@ -133,6 +133,7 @@
     <Compile Include="Plugins\ExtensionManifestTests.cs" />
     <Compile Include="ProcessStarterTests.cs" />
     <Compile Include="SigningToolsTests.cs" />
+    <Compile Include="SortableNameConverterTests.cs" />
     <Compile Include="TestPlayniteApplication.cs" />
     <Compile Include="XmlTests.cs" />
     <Compile Include="SizesTests.cs" />
